5th commit 한 날짜의 마지막 입력된 점수만 입력되도록 수정 테스트의 오류(타입 불일치, 객체 미생성, 설계 변경 미반영, null 처리 미고려) 수정

+++ src/main/java/com/staredu/grammar/repository/SpringDataJpaScoreRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SpringDataJpaScoreRepository extends JpaRepository<Score, Integer>, ScoreRepository {
+public interface SpringDataJpaScoreRepository extends JpaRepository<Score, Long>, ScoreRepository {
 
     Optional<Score> findByScore(Integer score);
 }
